Choosing correct precipitation variable from RDRS
Previously, the re-forecast precipitation variable was chosen to be fed
to the Alberta Agriculture Irrigation Demand Model (IDM). This has been
an error from the author's side, and this commit aims to correct it. In
this commit, the correct CaPA output of the RDRSv2.1 dataset has been
selected for further analysis.

The incorrect RDRSv2.1 precipitation variable was: 'RDRS_v2.1_P_PR0_SFC'
The correct RDRSv2.1 precipitation variable is: 'RDRS_v2.1_A_PR0_SFC'

Apologies to the Province of Alberta for this error from University of
Calgary's side.

Reported-by: Prabin Rokaya <prabin.rokaya@gov.ab.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>